rename a component
